+++ src/test/java/com/ashisht/mini_bank/controller/TransactionControllerTest.java
@@ -12,8 +12,6 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.math.BigDecimal;
-
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -34,7 +32,7 @@ class TransactionControllerTest {
         transaction.setTransactionId(1L);
         transaction.setAccount(Account.builder().accountId(1L).build());
         transaction.setOperationType(OperationType.builder().operationTypeId(1).description("CASH PURCHASE").build());
-        when(transactionService.createTransaction(anyLong(), anyInt(), any(BigDecimal.class))).thenReturn(transaction);
+        when(transactionService.createTransaction(anyLong(), anyInt(), any(Double.class))).thenReturn(transaction);
 
         String requestBody = "{\"accountId\":1,\"operationTypeId\":1,\"amount\":100.0}";
         mockMvc.perform(post("/transactions")
@@ -47,7 +45,7 @@ class TransactionControllerTest {
     @Test
     @DisplayName("POST /transactions - invalid input")
     void createTransaction_invalidInput() throws Exception {
-        when(transactionService.createTransaction(anyLong(), eq(999), any(BigDecimal.class)))
+        when(transactionService.createTransaction(anyLong(), eq(999), any(Double.class)))
                 .thenThrow(new IllegalArgumentException("Operation type not found"));
 
         String requestBody = "{\"accountId\":1,\"operationTypeId\":999,\"amount\":100.0}";
